Fix race condition in groupme worker
Because the match and decline match controller action grab a lock on the item in question, if the ResultsGroupmeWorker starts processing prior to the lock being released, the old version of the item gets used to generate the groupme message. This led to "Decline Match" actions looking like regular bid reports. By grabbing the lock in the worker, we guarantee that we have the most up to date version before sending.